requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0

# Azure OpenAI SDK
openai==1.14.2

# Data & Modeling
pydantic==2.6.4
numpy==1.26.4
pydantic-settings==2.2.1


# HTTP Client
httpx==0.27.0

# Vector Store (FAISS)
faiss-cpu==1.7.4

# PDF Generation
reportlab==4.1.0

# Utility
python-dotenv==1.0.1

# Optional but recommended (for future extensions)
orjson==3.10.0


